add pillow and lz4 as required dependencies
Multiple times people have pip installed deppth2 for hades and did not have Pillow or LZ4 installed - and did not know why deppth was not working.
Although it says to install them in cmd line and in readme, it would be easier to bundle the 3 as one for new users.